dom/tests/mochitest/sessionstorage/test_sessionStorageBase.html

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>sessionStorage basic test</title>

<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">

var expectedEvents = [
  "empty,null,",
  "empty,,null",
  "key1,null,value1",
  "key1,value1,null",
  "key1,null,value1",
  "key2,null,value2",
  "key2,value2,value2-2",
  "key1,value1,value1-2",
  "key2,value2-2,null",
  "null,null,null"
];

function setup() {
  sessionStorage.clear();
  SimpleTest.executeSoon(startTest);
}

function startTest()
{
  // Initially check the sessionStorage is empty
  is(sessionStorage.length, 0, "The storage is empty [1]");
  is(sessionStorage.key(0), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.key(-1), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.key(1), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.getItem("nonexisting"), null, "Nonexisting item is null (getItem())");
  is(sessionStorage.nonexisting, undefined, "Nonexisting item is undefined (array access)");
  is(sessionStorage.nonexisting, undefined, "Nonexisting item is undefined (property access)");
  sessionStorage.removeItem("nonexisting"); // Just check there is no exception

  is(typeof sessionStorage.getItem("nonexisting"), "object", "getItem('nonexisting') is object");
  is(typeof sessionStorage.nonexisting, "undefined", "['nonexisting'] is undefined");
  is(typeof sessionStorage.nonexisting, "undefined", "nonexisting is undefined");
  is(typeof sessionStorage.getItem("nonexisting2"), "object", "getItem('nonexisting2') is object");
  is(typeof sessionStorage.nonexisting2, "undefined", "['nonexisting2'] is undefined");
  is(typeof sessionStorage.nonexisting2, "undefined", "nonexisting2 is undefined");

  var sessionStorageCopy = sessionStorage;

  function onStorageChanged(e) {
    if (e.storageArea == sessionStorageCopy) {
      ok(expectedEvents.length, "Not more then expected events encountered");
      var receivedEvent = e.key + "," + e.oldValue + "," + e.newValue;
      is(receivedEvent, expectedEvents.shift(), "Expected event data: " + receivedEvent);
    }
  }

  // Listen for MozSessionStorageChanged
  SpecialPowers.addChromeEventListener("MozSessionStorageChanged", onStorageChanged, true);

  // add an empty-value key
  sessionStorage.setItem("empty", "");
  is(sessionStorage.getItem("empty"), "", "Empty value (getItem())");
  is(sessionStorage.empty, "", "Empty value (array access)");
  is(sessionStorage.empty, "", "Empty value (property access)");
  is(typeof sessionStorage.getItem("empty"), "string", "getItem('empty') is string");
  is(typeof sessionStorage.empty, "string", "['empty'] is string");
  is(typeof sessionStorage.empty, "string", "empty is string");
  sessionStorage.removeItem("empty");
  is(sessionStorage.length, 0, "The storage has no keys");
  is(sessionStorage.getItem("empty"), null, "empty item is null (getItem())");
  is(sessionStorage.empty, undefined, "empty item is undefined (array access)");
  is(sessionStorage.empty, undefined, "empty item is undefined (property access)");
  is(typeof sessionStorage.getItem("empty"), "object", "getItem('empty') is object");
  is(typeof sessionStorage.empty, "undefined", "['empty'] is undefined");
  is(typeof sessionStorage.empty, "undefined", "empty is undefined");

  // add one key, check it is there
  sessionStorage.setItem("key1", "value1");
  is(sessionStorage.length, 1, "The storage has one key-value pair");
  is(sessionStorage.key(0), "key1");
  is(sessionStorage.key(-1), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.key(1), null, "key() should return null for out-of-bounds access");

  // check all access method give the correct result
  // and are of the correct type
  is(sessionStorage.getItem("key1"), "value1", "getItem('key1') == value1");
  is(sessionStorage.key1, "value1", "['key1'] == value1");
  is(sessionStorage.key1, "value1", "key1 == value1");

  is(typeof sessionStorage.getItem("key1"), "string", "getItem('key1') is string");
  is(typeof sessionStorage.key1, "string", "['key1'] is string");
  is(typeof sessionStorage.key1, "string", "key1 is string");

  // remove the previously added key and check the storage is empty
  sessionStorage.removeItem("key1");
  is(sessionStorage.length, 0, "The storage is empty [2]");
  is(sessionStorage.key(0), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.getItem("key1"), null, "\'key1\' removed");

  is(typeof sessionStorage.getItem("key1"), "object", "getItem('key1') is object");
  is(typeof sessionStorage.key1, "undefined", "['key1'] is undefined");
  is(typeof sessionStorage.key1, "undefined", "key1 is undefined");

  // add one key, check it is there
  sessionStorage.setItem("key1", "value1");
  is(sessionStorage.length, 1, "The storage has one key-value pair");
  is(sessionStorage.key(0), "key1");
  is(sessionStorage.getItem("key1"), "value1");

  // add a second key
  sessionStorage.setItem("key2", "value2");
  is(sessionStorage.length, 2, "The storage has two key-value pairs");
  is(sessionStorage.getItem("key1"), "value1");
  is(sessionStorage.getItem("key2"), "value2");
  var firstKey = sessionStorage.key(0);
  var secondKey = sessionStorage.key(1);
  ok((firstKey == 'key1' && secondKey == 'key2') ||
     (firstKey == 'key2' && secondKey == 'key1'),
     'key() API works.');

  // change the second key
  sessionStorage.setItem("key2", "value2-2");
  is(sessionStorage.length, 2, "The storage has two key-value pairs");
  is(sessionStorage.key(0), firstKey); // After key value changes the order must be preserved
  is(sessionStorage.key(1), secondKey);
  is(sessionStorage.key(-1), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.key(2), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.getItem("key1"), "value1");
  is(sessionStorage.getItem("key2"), "value2-2");

  // change the first key
  sessionStorage.setItem("key1", "value1-2");
  is(sessionStorage.length, 2, "The storage has two key-value pairs");
  is(sessionStorage.key(0), firstKey); // After key value changes the order must be preserved
  is(sessionStorage.key(1), secondKey);
  is(sessionStorage.key(-1), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.key(2), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.getItem("key1"), "value1-2");
  is(sessionStorage.getItem("key2"), "value2-2");

  // remove the second key
  sessionStorage.removeItem("key2");
  is(sessionStorage.length, 1, "The storage has one key-value pair");
  is(sessionStorage.key(0), "key1");
  is(sessionStorage.key(-1), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.key(1), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.getItem("key1"), "value1-2");

  // Clear the storage
  sessionStorage.clear();
  is(sessionStorage.length, 0, "The storage is empty [3]");
  is(sessionStorage.key(0), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.key(-1), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.key(1), null, "key() should return null for out-of-bounds access");
  is(sessionStorage.getItem("nonexisting"), null, "Nonexisting item is null");
  is(sessionStorage.getItem("key1"), null, "key1 removed");
  is(sessionStorage.getItem("key2"), null, "key2 removed");
  sessionStorage.removeItem("nonexisting"); // Just check there is no exception
  sessionStorage.removeItem("key1"); // Just check there is no exception
  sessionStorage.removeItem("key2"); // Just check there is no exception

  SimpleTest.executeSoon(function () {
    SpecialPowers.removeChromeEventListener("MozSessionStorageChanged", onStorageChanged, true);
    is(expectedEvents.length, 0, "received the correct number of events");

    sessionStorage.clear();
    SimpleTest.finish();
  });
}

SimpleTest.waitForExplicitFinish();

</script>

</head>

<body onload="setup();">

</body>
</html>
